Updated credit on config files.
After Zygon telling me the credits were missing i realized these files were actually quickly grabbed from desu when i needed things i didn't care about to load very quickly. This should be the last bit of things that were accidentally not credited. When creating a file in netbeans it automatically adds my GNU license information. Therefore when i paste in a chunk of code it's easy to forget about the credit. This marks a total of 4 files not written by me that weren't credited. I sincerely apologise for this lack of proper crediting and will keep a better eye on it in the future before making repos public.,